Crate fe_common

Re-exports

pub use files::File;

pub use files::FileKind;

pub use files::SourceFileId;

Modules

db

diagnostics

files

numeric

panic

utils

Macros

assert_snapshot_wasm

assert_strings_eq

Compare the given strings and panic when not equal with a colorized line diff.

impl_intern_key

Structs

Span

An exclusive span of byte offsets in a source file.

Traits

Spanned